backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
chromadb==0.4.18
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
# LLM和嵌入模型依赖
dashscope>=1.14.0
# 文档处理依赖 - 使用兼容版本
langchain-text-splitters>=0.2.0
langchain-core>=0.2.0
# NumPy版本兼容性
numpy<2.0
# 语义分块依赖
scikit-learn>=1.3.0
nltk>=3.8.0
jieba>=0.42.1
